requirements.txt
numpy
opencv-python
h5py
pillow
pympler